add tests for new dataset class
